Switch from H2 to PostgreSQL database
Updated the database configuration to use PostgreSQL instead of H2 by modifying the schema.sql and application-dev.properties files. Removed H2 dependency from pom.xml and added PostgreSQL and Docker Compose dependencies to support the new setup. Adjusted schema to ensure it creates the table only if it doesn't already exist and included a unique constraint on the username field.